build: move from doozer to travis
exec: parser: readline: fix some warnings
build: force minimum boost version

+++ src/readline/terminal.cc
@@ -1,6 +1,8 @@
 #include "terminal.hh"
+#include <cassert>
 #include <termios.h>
 #include <unistd.h>
+#include <iostream>
 
 namespace readline
 {
@@ -10,7 +12,7 @@ TerminalCfg::TerminalCfg()
   tcgetattr(STDIN_FILENO, &*term_config_);
 
   auto new_config = *term_config_;
-  new_config.c_lflag &= ~(ICANON | ECHO);
+  new_config.c_lflag &= static_cast<unsigned int>(~(ICANON | ECHO));
   tcsetattr(STDIN_FILENO, TCSANOW, &new_config);
 }
